+++ packages/block-library/src/embed/embed-loading.js
@@ -1,0 +1,14 @@
+/**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { Spinner } from '@wordpress/components';
+
+const EmbedLoading = () => (
+	<div className="wp-block-embed is-loading">
+		<Spinner />
+		<p>{ __( 'Embedding…' ) }</p>
+	</div>
+);
+
+export default EmbedLoading;
